refactor(phase4): manual presentation層をmanual/presentationに統合
