--- lib/features/manual_editing/presentation/presentation.dart
@@ -1,6 +1,0 @@
-// Presentation layer exports for manual editing feature
-export 'providers/manual_edit_providers.dart';
-export 'screens/manual_edit_screen.dart';
-export 'widgets/manual_header_widget.dart';
-export 'widgets/step_list_widget.dart';
-export 'widgets/step_edit_dialog.dart';
